SWC-6990
Fix import cycle
