+++ packages/synapse-react-client/src/components/CardContainerLogic/index.ts
@@ -7,4 +7,4 @@ export type {
   ColumnSpecifiedLink,
   CardContainerLogicProps,
 } from './CardContainerLogic'
-export { default, CardContainerLogic, TargetEnum } from './CardContainerLogic'
+export { default, CardContainerLogic } from './CardContainerLogic'
