+++ packages/synapse-react-client/src/components/GenericCard/TableRowGenericCard.vitest.tsx
@@ -26,7 +26,7 @@ import { act, fireEvent, render, screen, within } from '@testing-library/react'
 import { cloneDeep } from 'lodash-es'
 import { mockAllIsIntersecting } from 'react-intersection-observer/test-utils'
 import { beforeEach, describe, it, test, vi } from 'vitest'
-import { TargetEnum } from '../CardContainerLogic'
+import { TargetEnum } from '@/utils/html/TargetEnum'
 import { EntityDownloadConfirmation } from '../EntityDownloadConfirmation'
 import { QueryVisualizationWrapper } from '../QueryVisualizationWrapper'
 import QueryWrapper from '../QueryWrapper'
