+++ packages/synapse-react-client/src/utils/html/TargetEnum.ts
@@ -1,0 +1,6 @@
+export enum TargetEnum {
+  CURRENT_WINDOW = '_self',
+  NEW_WINDOW = '_blank',
+  PARENT_FRAME = '_parent',
+  FULL_WINDOW_BODY = '_top',
+}
